Fix up build rules to enable Kotlin support
The app-compat-annotations library doesn't play nicely with the Kotlin
compiler. This change isolates the files that rely on this into their
own build rules and bypass the DocumentsUI-lib rule which will contain
Kotlin files in the future.

This change is almost identical to the change done in Settings at
http://ag/18655752.

Bug: 370880067
Test: mma DocumentsUIGoogle
Flag: EXEMPT only build file changes
Change-Id: Id9f2e6dce46f5251f7a99c12d94f50c68597dd19