[DocsUI M3] Add compose folder to experiment compose
Add a new folder "compose" inside DocumentsUI so we can have a
isolate environment to experiment constructing Jetpack Compose
components. It's independent from the original DocumentsUI and
has its own build target. We could use this as a Compose Lab to
build component here first and then reuse it in the original
DocumentsUI.

Bug: 374224390
Test: m DocumentsUICompose
Flag: EXEMPT experiment for dev only
Change-Id: I33d61d47889e16c39aa92de082c2fdba2193461b