Respect is_launcher_enabled on PRE_BOOT_COMPLETED for FEATURE_PC
For devices running version S+ that undergo an update. The value for
is_launcher_enabled that is specified in the RRO is not respected.
Despite being read it is simply hardcoded to false. This means
that devices that have it enabled during development then undergo
an update it stays visible for a short period then promptly
disappears.

For now let's remove this behaviour when the FEATURE_PC system
feature is enabled.

Bug: 374866808
Test: m dist && update_device.py, verify DocsUI doesn't hide
Flag: EXEMPT desktop-only change
Change-Id: I776c59dd4421566aab3cf122f2cb3dbed9a45164