[DocsUI Search] Add a DocsUI search flag.
The search flag is used to control whether the current or the next
generation of search functionality is used in DocumentsUI. The flag
will be used in both Java code and XML files.
Adding use_ to make the if statement more natural.

Test: m DocumentsUIGoogle
Bug: 378590312
Flag: com.android.documentsui.flags.use_search_v2

Change-Id: Idf1fd4486245a44cfbdbce6733620e3765ae09a0