DocsUI M3: Remove the variant 720dp-land.
For M3 we're moving away from variances based on the height of the
window (landscape vs portrait), instead it'll based only on the window
width.

Move the layouts.xml to 720dp so it continue to use the fixed_layout for
windows larger than 720dp wide.

Test: Manual build, deploy apk and start it on DUT with flag on
Flag: com.android.documentsui.flags.use_material3
Bug: 380369329
Change-Id: I27f757ebbcf2302f905efafd9b8c8b939f2e15f4